Bug Fix n Refactoring
0. Player Pointers getting inspector n calling Pointers at player instance .
1. Bridge Bug Fix.
2. Add Lift Up n Down Object.
3. Lift Script Fixed.
4. Player Prefabs Fixed.
5. etc...
