+++ WAGTAIL/Assets/01_Scripts/02_Object/InteractionSystem/PickUp.cs
@@ -10,6 +10,11 @@ public class PickUp : MonoBehaviour, IInteractable
 
     public string InteractionPrompt => _promt;
 
+    private void Start()
+    {
+        _playerEquipPoint = Player.Instance.EquipPoint.gameObject;  
+    }
+
     public bool Interact(Interactor interactor)
     {
         // TODO : Player가 nomal상태인지 확인해주는거 구현해야됨
