+++ WAGTAIL/Assets/01_Scripts/02_Object/InteractionSystem/Push.cs
@@ -16,6 +16,11 @@ public class Push : MonoBehaviour, IInteractable
 
     public string InteractionPrompt => _promt;
 
+    private void Start()
+    {
+        _playerEquipPoint = Player.Instance.EquipPoint.gameObject;
+    }
+
     public bool Interact(Interactor interactor)
     {
         if (interactor.player.movementSM.currentState == interactor.player.idle)
