+++ WAGTAIL/Assets/02_Prefabs/02_Object/LiftObj.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e40e9507daf01fa4fbda2a7121c7a446
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
